+++ spec/guide_section_sub_section_spec.rb
@@ -36,7 +36,7 @@ module DmtdVbmappData
     private
 
     def client
-      Client.retrieve_clients.first
+      Client.new(id: 57)
     end
   end