+++ spec/vbmapp_area_group_spec.rb
@@ -35,7 +35,7 @@ module DmtdVbmappData
     private
 
     def client
-      Client.retrieve_clients.first
+      Client.new(id: 57)
     end
 
     def group_index_json