+++ spec/vbmapp_area_question_spec.rb
@@ -52,7 +52,7 @@ module DmtdVbmappData
     private
 
     def client
-      Client.retrieve_clients.first
+      Client.new(id: 57)
     end
 
     def question_json