+++ spec/vbmapp_area_spec.rb
@@ -25,7 +25,7 @@ module DmtdVbmappData
     private
 
     def client
-      Client.retrieve_clients.first
+      Client.new(id: 57)
     end
 
     def area_index_json